+++ src/kraken/ingests.rs
@@ -15,9 +15,6 @@
 // (Modifications|Other (data|code)|Everything else) Copyright 2019 the
 // libtwitch-rs authors.  See copying.md for further legal info.
 
-extern crate chrono;
-extern crate serde_json;
-
 use super::super::{
 	response::TwitchResult,
 	TwitchClient,