+++ src/kraken/teams.rs
@@ -15,16 +15,12 @@
 // (Modifications|Other (data|code)|Everything else) Copyright 2019 the
 // libtwitch-rs authors.  See copying.md for further legal info.
 
-extern crate chrono;
-extern crate serde_json;
-extern crate urlparse;
-
 use std::{
 	self,
 	io::Write,
 };
 
-use self::chrono::prelude::*;
+use chrono::prelude::*;
 use serde::Deserialize;
 
 use super::users::User;
@@ -37,7 +33,7 @@ use crate::{
 /// Gets all active teams
 ///
 /// #### Authentication: `None`
-pub fn get_all<'c>(c: &'c TwitchClient) -> TwitchResult<TeamIterator<'c>> {
+pub fn get_all(c: &TwitchClient) -> TwitchResult<TeamIterator> {
 	let iter = TeamIterator {
 		client: c,
 		cur: None,